+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 [Full Changelog](https://github.com/ZelionGG/KeystonePercentageHelper/compare/1.7...1.7.1) [Previous Releases](https://github.com/ZelionGG/KeystonePercentageHelper/releases)
 
 - **Season 3** is now fully operational, Eco-Dome Al'dani should show the correct name and icon. Enjoy Season 3 preparation!
+- Korean translation updated, thank you again **BlueSea-jun**.
+- Russian translation updated, thank you **Hollicsh**.
+